Redelete tests
